requirements.txt
tensorflow
opencv-python
pycocotools
pillow
tf-models-official
